requirements.txt
icrawler
pyfiglet
Pillow
